+++ CHANGES.md
@@ -4,11 +4,23 @@ Release Notes.
 
 0.1.0
 ------------------
-#### Project
-
 #### Features
+* Build the Satellite core structure.
+* Add prometheus self telemetry.
+* Add kafka client plugin.
+* Add none-fallbacker plugin.
+* Add timer-fallbacker plugin.
+* Add nativelog-kafka-forwarder plugin.
+* Add memory-queue plugin.
+* Add mmap-queue plugin.
+* Add grpc-nativelog-receiver plugin.
+* Add http-nativelog-receiver plugin.
+* Add grpc-server plugin.
+* Add http-server plugin.
+* Add prometheus-server plugin.
 
 #### Bug Fixes
 
-All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/64?closed=1)
-
+#### Issues and PR
+- All issues  are [here](https://github.com/apache/skywalking/milestone/64?closed=1)  
+- All and pull requests are [here](https://github.com/apache/skywalking-satellite/pulls?q=is%3Aopen+is%3Apr+milestone%3A0.1.0)